organize code with utils.js, rename getDefaultProblem() to getDefaultLink()
