+++ AWS/lambdafunc.js
@@ -1,8 +1,8 @@
 const AWS = require("aws-sdk");
-var XMLHttpRequest = require('xhr2');
 AWS.config.update({ 
     region: "us-west-1"
 });
+const { buildResponse, scanDynamo, getDifficulty, getLatestProblem, getDefaultLink } = require('./utils.js');
 
 const DYANMODB = new AWS.DynamoDB.DocumentClient();
 const TABLE_NAME = "leetcode_tracker";
@@ -52,7 +52,6 @@ async function createUser(userId, username) {
         TotalCount: 0
     }
 
-    // Add to DynamoDB
     const params = {
         TableName: TABLE_NAME,
         Item: newUser,
@@ -61,6 +60,7 @@ async function createUser(userId, username) {
         },
         ConditionExpression: `attribute_not_exists(userId)`
     }
+    
     return await DYANMODB.put(params).promise().then(() => {
         const body = {
             Operation: 'SAVE',
@@ -80,11 +80,8 @@ async function createUser(userId, username) {
 
 async function updateUser(requestBody) {
     const userId = requestBody['userId'];
-    const URL = getDefaultProblem(requestBody['link']);
+    const URL = getDefaultLink(requestBody['link']);
 
-    /**
-     * Setup problemObj with userId, URL, & difficulty
-     */
     let problemObj = {
         link: URL,
         date: new Date().toLocaleString("en-US", { timeZone: "America/Los_Angeles" }),
@@ -195,7 +192,7 @@ async function getRanking(difficulty) {
         TableName: TABLE_NAME
     }
     
-    const top3 = await scanDynamo(params, []).then((allUsers) => {
+    const top3 = await scanDynamo(DYANMODB, params, []).then((allUsers) => {
         return allUsers.sort((a, b) => b[category] - a[category]);  // Sort in Descending Order
     });
 
@@ -206,122 +203,4 @@ async function getRanking(difficulty) {
     }
   
     return buildResponse(200, ranking);
-}
-
-function buildResponse(statusCode, body) {
-    return {
-        statusCode: statusCode,
-        headers: {
-            'Content-Type': 'application/json',
-            "Access-Control-Allow-Headers" : "Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token",
-            "Access-Control-Allow-Methods" : "OPTIONS,POST,PUT,GET,DELETE,PATCH",
-            "Access-Control-Allow-Credentials" : true,
-            "Access-Control-Allow-Origin" : "*",
-            "X-Requested-With" : "*"
-        },
-        body: JSON.stringify(body)
-    }
-}
-
-async function scanDynamo(scanParams, itemArray) {
-    try {
-      const dynamoData = await DYANMODB.scan(scanParams).promise();
-      itemArray = itemArray.concat(dynamoData.Items);
-      if (dynamoData.LastEvaluatedKey) {
-        scanParams.ExclusiveStartkey = dynamoData.LastEvaluatedKey;
-        return await scanDynamo(scanParams, itemArray);
-      }
-      return itemArray;
-    } catch(error) {
-      console.error('Scan Error, could not scan! : ', error);
-    }
-}
-
-/**
- * Functions to get the difficulty of a problem given the link
- */
-
-// Source: https://stackoverflow.com/questions/247483/http-get-request-in-javascript 
-function httpGetAsync(URL) {
-  return new Promise((resolve, reject) => {
-    let xmlHttp = new XMLHttpRequest();
-    xmlHttp.onreadystatechange = function() { 
-        if (xmlHttp.readyState == 4 && xmlHttp.status == 200) {
-            resolve(xmlHttp.responseText);
-        }
-    }
-    xmlHttp.open("GET", URL, true); // true for asynchronous 
-    xmlHttp.send(null); 
-  });
-}
-
-async function getDifficulty(URL) {
-    const text = await httpGetAsync(URL);
-
-    // Isolate the value for Difficulty: Easy/Medium/Hard
-    const findDifficulty = text.split('"difficulty"')[1];
-
-    // Remove the rest of the json data
-    const isolateDifficulty = findDifficulty.split(',')[0];
-    
-    // Remove the double quotations & semicolon with regex
-    const difficulty = isolateDifficulty.replace(/['":]+/g, '');
-
-    // Return the difficulty
-    return difficulty;
-}
-
-function getLatestProblem(problems) {
-    // If no problems completed yet return problem with empty nest emojis
-    if (problems.length === 0) {
-        return {
-            'link': '🪹',
-            'date': '🪹',
-            'difficulty': '🪹',
-        }
-    }
-
-    // Start comparing from the zero-th index. Dates can only compare to other dates
-    let latestDate = problems[0].date;
-    problems.map((problem) => {
-        if (problem.date > latestDate) {
-            latestDate = problem.date;
-        }
-    });
-
-    // Since we know the latest date in the problems array, find the problem matching the time
-    const latestProblem = problems.find(problem => {
-        return problem.date === latestDate.toLocaleString()
-    });
-
-    return latestProblem;
-}
-
-function getDefaultProblem(url) {
-    /**
-     * The purpose of this function is to get the "default link". I didn't want to find the regex for it. Ex:
-     * 
-     * Default:     https://leetcode.com/problems/two-sum/
-     * Not Default: https://leetcode.com/problems/two-sum/discussion/
-     */
-
-    /**
-     * splitBySlash looks like:
-     * 
-     * [
-        'https:',
-        '',
-        'leetcode.com',
-        'problems',
-        'two-sum',
-        'discussion',   // Assuming there is an extra tag
-        ''
-     * ]
-     */
-    const splitBySlash = url.split('/');
-
-    // Combined together: https: + "//" + leetcode.com + "/" + problems + "/" + two-sum + "/"
-    let defaultUrl = splitBySlash[0] + "//" + splitBySlash[2] + "/" + splitBySlash[3] + "/" + splitBySlash[4] + "/";
-
-    return defaultUrl;
 }
